requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
confuse==1.4.0
dlib==19.22.0
face-recognition==1.3.0
face-recognition-models==0.3.0
idna==2.10
numpy==1.20.2
opencv-python==4.5.1.48
paho-mqtt==1.5.1
Pillow==8.2.0
PyYAML==5.4.1
requests==2.25.1
urllib3==1.26.4
